Removes old useless driver entrypoints
When W2K was initially being tested at the very beginning, I put basic
entrypoints into the drivers I was working on support for which would
send a basic message to the logs preaching the version of Minecraft they
were intended for, mainly so that I could see which drivers loaded and
which ones didn't and therefore determine if the model I was going for
was a good idea.

It worked, but I never removed those entrypoints until now and they
serve very little purpose outside of the testing I was doing and it
could have caused a bunch of confusion somewhere down the road, so I'm
removing them now that their purpose has been fulfilled and there is no
reason to keep them anymore.