Don't integrate with mods if not loaded by Fabric
The previous commit offered the ability to integrate into certain mods
that were loaded with non-Fabric loaders using mods like Fabricated
Rift. I was optimistic that it would be good for integrating mods in
specific environments (like running Litematica on 1.13.2, for example)
and while initial testing proved promising, it was just too good to be
true at the end of the day.

This commit removes the ability to integrate into non-Fabric mods while
also retaining the non-intrusive Litematica integration that was added
in the last commit.